Move tracing and profiling component building out of the core namespace
